fix: move menu item to common components gf-88
