--- apps/frontend/src/libs/components/menu/libs/components/components.ts
@@ -1,1 +1,0 @@
-export { MenuItem } from "./menu-item/menu-item.js";
